+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/webview/in_app_webview/InAppWebViewSettings.java
@@ -87,8 +87,10 @@ public class InAppWebViewSettings implements ISettings<InAppWebViewInterface> {
   public Integer disabledActionModeMenuItems;
   public String fantasyFontFamily = "fantasy";
   public String fixedFontFamily = "monospace";
-  public Integer forceDark = 0; // WebSettingsCompat.FORCE_DARK_OFF
-  public Integer forceDarkStrategy = WebSettingsCompat.DARK_STRATEGY_PREFER_WEB_THEME_OVER_USER_AGENT_DARKENING;
+  @Nullable @Deprecated
+  public Integer forceDark = null;
+  @Nullable @Deprecated
+  public Integer forceDarkStrategy = null;
   public Boolean geolocationEnabled = true;
   public WebSettings.LayoutAlgorithm layoutAlgorithm;
   public Boolean loadWithOverviewMode = true;
@@ -106,11 +108,14 @@ public class InAppWebViewSettings implements ISettings<InAppWebViewInterface> {
   public Boolean supportMultipleWindows = false;
   public String regexToCancelSubFramesLoading;
   public Integer overScrollMode = View.OVER_SCROLL_IF_CONTENT_SCROLLS;
+  @Nullable
   public Boolean networkAvailable = null;
   public Integer scrollBarStyle = View.SCROLLBARS_INSIDE_OVERLAY;
   public Integer verticalScrollbarPosition = View.SCROLLBAR_POSITION_DEFAULT;
+  @Nullable
   public Integer scrollBarDefaultDelayBeforeFade = null;
   public Boolean scrollbarFadingEnabled = true;
+  @Nullable
   public Integer scrollBarFadeDuration = null;
   @Nullable
   public Map<String, Object> rendererPriorityPolicy = null;